# crates/sdkwork-routes-notary-http-auth/src/response.rs
use axum::{
    http::{HeaderName, HeaderValue, StatusCode},
    response::{IntoResponse, Response},
    Json,
};
use sdkwork_utils_rust::{SdkWorkApiResponse, SdkWorkProblemDetail, SdkWorkResultCode};
use sdkwork_web_core::WebRequestContext;
use serde_json::Value;

use crate::NotaryAuthError;

#[derive(Clone, Debug, Eq, PartialEq)]
pub struct NotaryRouteError {
    pub status: StatusCode,
    pub result_code: SdkWorkResultCode,
    pub message: String,
}

impl From<NotaryAuthError> for NotaryRouteError {
    fn from(error: NotaryAuthError) -> Self {
        Self {
            status: error.status,
            result_code: SdkWorkResultCode::AuthenticationRequired,
            message: error.message,
        }
    }
}

impl NotaryRouteError {
    pub fn invalid_parameter(message: impl Into<String>) -> Self {
        Self {
            status: StatusCode::BAD_REQUEST,
            result_code: SdkWorkResultCode::InvalidParameter,
            message: message.into(),
        }
    }

    pub fn from_wire(status: StatusCode, wire_code: &str, message: impl Into<String>) -> Self {
        Self {
            status,
            result_code: result_code_from_wire(wire_code),
            message: message.into(),
        }
    }

    pub fn into_response_for(self, ctx: &WebRequestContext) -> Response {
        let trace_id = ctx.resolved_trace_id();
        let problem = SdkWorkProblemDetail::platform(self.result_code, self.message, trace_id);
        let mut response = (
            self.status,
            [(axum::http::header::CONTENT_TYPE, "application/problem+json")],
            Json(problem),
        )
            .into_response();
        attach_trace_header(&mut response, &ctx.resolved_trace_id());
        response
    }
}

impl IntoResponse for NotaryRouteError {
    fn into_response(self) -> Response {
        let trace_id = uuid_fallback_trace_id();
        let problem = SdkWorkProblemDetail::platform(self.result_code, self.message, trace_id);
        (
            self.status,
            [(axum::http::header::CONTENT_TYPE, "application/problem+json")],
            Json(problem),
        )
            .into_response()
    }
}

pub fn success_envelope(ctx: &WebRequestContext, status: StatusCode, data: Value) -> Response {
    let trace_id = ctx.resolved_trace_id();
    let envelope = SdkWorkApiResponse::success(data, trace_id.clone());
    let mut response = (status, Json(envelope)).into_response();
    attach_trace_header(&mut response, &trace_id);
    response
}

pub fn finish_success(
    ctx: &WebRequestContext,
    status: StatusCode,
    data: Value,
) -> Result<Response, NotaryRouteError> {
    Ok(success_envelope(ctx, status, data))
}

pub fn finish_success_no_content(ctx: &WebRequestContext) -> Result<Response, NotaryRouteError> {
    let trace_id = ctx.resolved_trace_id();
    let mut response = StatusCode::NO_CONTENT.into_response();
    attach_trace_header(&mut response, &trace_id);
    Ok(response)
}

const NOTARY_APP_CREATED_OPERATIONS: &[&str] = &[
    "notary.cases.create",
    "notary.cases.acceptances.create",
    "notary.cases.rejections.create",
    "notary.cases.completions.create",
    "notary.cases.parties.create",
    "notary.cases.parties.signatures.create",
    "notary.cases.parties.videoInvites.create",
    "notary.cases.parties.signatureInvites.create",
    "notary.cases.files.create",
    "notary.cases.downloadPackages.create",
    "notary.cases.assignments.create",
];

const NOTARY_BACKEND_CREATED_OPERATIONS: &[&str] = &[
    "notary.organizationProfiles.create",
    "notary.matters.create",
    "notary.cases.assignments.create",
];

pub fn success_status_for_notary_app_operation(operation_id: &str) -> StatusCode {
    if NOTARY_APP_CREATED_OPERATIONS.contains(&operation_id) {
        StatusCode::CREATED
    } else {
        StatusCode::OK
    }
}

pub fn success_status_for_notary_backend_operation(operation_id: &str) -> StatusCode {
    if NOTARY_BACKEND_CREATED_OPERATIONS.contains(&operation_id) {
        StatusCode::CREATED
    } else {
        StatusCode::OK
    }
}

fn attach_trace_header(response: &mut Response, trace_id: &str) {
    if let Ok(value) = HeaderValue::from_str(trace_id) {
        response
            .headers_mut()
            .insert(HeaderName::from_static("x-sdkwork-trace-id"), value);
    }
}

fn result_code_from_wire(wire_code: &str) -> SdkWorkResultCode {
    match wire_code {
        "unauthenticated" => SdkWorkResultCode::AuthenticationRequired,
        "unauthorized" => SdkWorkResultCode::PermissionRequired,
        "not-found" => SdkWorkResultCode::NotFound,
        "conflict" | "invalid-state" => SdkWorkResultCode::Conflict,
        "validation" => SdkWorkResultCode::ValidationError,
        "transport" => SdkWorkResultCode::BadGateway,
        "provider-unavailable" => SdkWorkResultCode::ServiceUnavailable,
        "storage" | "unknown" => SdkWorkResultCode::InternalError,
        _ => SdkWorkResultCode::InternalError,
    }
}

fn uuid_fallback_trace_id() -> String {
    sdkwork_utils_rust::uuid()
}

#[cfg(test)]
mod tests {
    use super::*;
    use axum::body::to_bytes;
    use sdkwork_web_core::{ServerRequestId, WebApiSurface, WebAuthMode, WebTransportFacts};

    fn test_context() -> WebRequestContext {
        WebRequestContext {
            request_id: ServerRequestId("req-test".to_owned()),
            api_surface: WebApiSurface::AppApi,
            auth_mode: WebAuthMode::DualToken,
            principal: None,
            transport: WebTransportFacts {
                path: "/app/v3/api/notary/cases".to_owned(),
                method: "GET".to_owned(),
                auth_token_present: true,
                access_token_present: true,
                api_key_present: false,
                oauth_bearer_present: false,
                agent_token_present: false,
            },
            locale: None,
            client_kind: None,
            operation: None,
            trace_id: Some("trace-notary-test".to_owned()),
        }
    }

    #[tokio::test]
    async fn success_envelope_wraps_data_with_zero_code_and_trace_id() {
        let response = success_envelope(
            &test_context(),
            StatusCode::OK,
            serde_json::json!({ "item": { "caseId": "case-1" } }),
        );
        let body = to_bytes(response.into_body(), usize::MAX)
            .await
            .expect("body");
        let payload: Value = serde_json::from_slice(&body).expect("json");
        assert_eq!(0, payload["code"].as_i64().unwrap());
        assert_eq!("trace-notary-test", payload["traceId"].as_str().unwrap());
        assert_eq!(
            "case-1",
            payload["data"]["item"]["caseId"].as_str().unwrap()
        );
    }

    #[tokio::test]
    async fn route_error_uses_problem_json_with_numeric_code() {
        let response = NotaryRouteError::from_wire(
            StatusCode::FORBIDDEN,
            "unauthorized",
            "missing permission",
        )
        .into_response_for(&test_context());
        assert_eq!(
            "application/problem+json",
            response
                .headers()
                .get(axum::http::header::CONTENT_TYPE)
                .and_then(|value| value.to_str().ok())
                .unwrap_or_default()
        );
        let body = to_bytes(response.into_body(), usize::MAX)
            .await
            .expect("body");
        let payload: Value = serde_json::from_slice(&body).expect("json");
        assert_eq!(40301, payload["code"].as_i64().unwrap());
        assert_eq!("trace-notary-test", payload["traceId"].as_str().unwrap());
    }

    #[test]
    fn success_status_follows_openapi_contract() {
        assert_eq!(
            StatusCode::CREATED,
            success_status_for_notary_app_operation("notary.cases.create")
        );
        assert_eq!(
            StatusCode::CREATED,
            success_status_for_notary_app_operation("notary.cases.acceptances.create")
        );
        assert_eq!(
            StatusCode::CREATED,
            success_status_for_notary_app_operation("notary.cases.rejections.create")
        );
        assert_eq!(
            StatusCode::CREATED,
            success_status_for_notary_app_operation("notary.cases.completions.create")
        );
        assert_eq!(
            StatusCode::CREATED,
            success_status_for_notary_backend_operation("notary.organizationProfiles.create")
        );
        assert_eq!(
            StatusCode::OK,
            success_status_for_notary_backend_operation("notary.organizationProfiles.update")
        );
    }
}