# sdks/sdkwork-notary-app-sdk/openapi/notary-app-api.sdkgen.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Notary App API",
    "version": "0.1.0",
    "description": "User-facing notary API. Merchandise owns sellable SPU/SKU matter definitions, Order owns checkout and order-item state, Payment remains behind Order for payment execution, Drive owns files and folders, and IAM owns organization members."
  },
  "servers": [
    {
      "url": "/app/v3/api"
    }
  ],
...
  "paths": {
...
    "/app/v3/api/notary/cases": {
      "get": {
        "tags": [
          "notary"
        ],
        "operationId": "notary.cases.list",
        "summary": "List notary cases",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-notary",
        "x-sdkwork-api-authority": "sdkwork-notary.app",
        "x-sdkwork-domain": "notary",
        "x-sdkwork-resource": "notaryCase",
        "x-sdkwork-permission": "notary.cases.read",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-data-scope": "organization",
        "x-sdkwork-audit-event": "notary.cases.list",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Query"
          },
          {
            "$ref": "#/components/parameters/CaseStatus"
          },
          {
            "$ref": "#/components/parameters/SkuId"
          }
        ],
        "responses": {
          "200": {
            "description": "Notary case page",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/NotaryCasePage"
                        }
                      }
                    }
                  ]
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-source-route-crate": "sdkwork-routes-notary-app-api",
        "x-sdkwork-source": "crates/sdkwork-routes-notary-app-api/src/handlers.rs"
      },
      "post": {
        "tags": [
          "notary"
        ],
        "operationId": "notary.cases.create",
        "summary": "Create a notary case through Order checkout and a Drive notary folder",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-notary",
        "x-sdkwork-api-authority": "sdkwork-notary.app",
        "x-sdkwork-domain": "notary",
        "x-sdkwork-resource": "notaryCase",
        "x-sdkwork-permission": "notary.cases.create",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-data-scope": "organization",
        "x-sdkwork-audit-event": "notary.cases.create",
        "x-sdkwork-idempotent": true,
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotaryCaseRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Notary case",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/NotaryCase"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "400": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-source-route-crate": "sdkwork-routes-notary-app-api",
        "x-sdkwork-source": "crates/sdkwork-routes-notary-app-api/src/handlers.rs"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "SDKWork auth token carried as Authorization: Bearer <auth_token>."
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token",
        "description": "SDKWork access isolation token."
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128
        }
      },
...
      "PageSize": {
        "name": "page_size",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "Query": {
        "name": "q",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 100
        }
      },
      "CaseStatus": {
        "name": "status",
        "in": "query",
        "required": false,
        "schema": {
          "$ref": "#/components/schemas/NotaryCaseStatus"
        }
      },
      "SkuId": {
        "name": "sku_id",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
...
    },
...
    "schemas": {
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/SdkWorkPlatformErrorCode"
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int32",
            "minimum": 40011,
            "maximum": 40099
          }
        }
      },
      "PageInfo": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "mode"
        ],
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "offset",
              "cursor"
            ]
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          },
          "totalItems": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
...
      "NotaryCaseStatus": {
        "type": "string",
        "enum": [
          "PENDING_REVIEW",
          "PROCESSING",
          "COMPLETED",
          "REJECTED",
          "CANCELLED",
          "CREATE_FAILED"
        ]
      },
      "Party": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "role",
          "identityId"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "partyType": {
            "type": "string",
            "enum": [
              "natural_person",
              "legal_person",
              "organization"
            ]
          },
          "identityType": {
            "type": "string"
          },
          "identityId": {
            "type": "string",
            "description": "Masked identity number for display."
          },
          "identityNoLast4": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "address": {
            "type": "string"
          },
          "remarks": {
            "type": "string"
          },
          "identityValidDateStart": {
            "type": "string",
            "format": "date"
          },
          "identityValidDateEnd": {
            "type": "string",
            "format": "date"
          },
          "signatureUrl": {
            "type": "string"
          },
          "signatureNodeId": {
            "type": "string"
          },
          "verificationStatus": {
            "type": "string",
            "enum": [
              "pending",
              "verified",
              "failed",
              "expired"
            ]
          }
        }
      },
      "NotaryDocumentCategory": {
        "type": "string",
        "enum": [
          "identity",
          "evidence",
          "notary"
        ]
      },
      "NotaryDocument": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "size",
          "status",
          "category"
        ],
        "properties": {
          "nodeId": {
            "type": "string"
          },
          "driveSpaceId": {
            "type": "string"
          },
          "driveSpaceType": {
            "type": "string",
            "enum": [
              "notary"
            ]
          },
          "parentNodeId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "size": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "verified",
              "pending",
              "error"
            ]
          },
          "category": {
            "$ref": "#/components/schemas/NotaryDocumentCategory"
          },
          "materialCode": {
            "type": "string"
          },
          "partyId": {
            "type": "string"
          },
          "mediaResource": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "TimelineEvent": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "time",
          "event",
          "actor"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "time": {
            "type": "string",
            "format": "date-time"
          },
          "event": {
            "type": "string"
          },
          "actor": {
            "type": "string"
          },
          "eventType": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        }
      },
      "NotaryCase": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "caseNo",
          "createTime",
          "applicant",
          "title",
          "notary",
          "remarks",
          "type",
          "status",
          "fee",
          "orderId",
          "orderItemId",
          "skuId",
          "driveSpaceId",
          "driveSpaceType",
          "driveFolderNodeId",
          "parties",
          "documents",
          "timeline"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "caseNo": {
            "type": "string"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "processTime": {
            "type": "string",
            "format": "date-time"
          },
          "applicant": {
            "type": "string"
          },
          "applicantUserId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "notary": {
            "type": "string"
          },
          "primaryNotaryMembershipId": {
            "type": "string"
          },
          "remarks": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/NotaryCaseStatus"
          },
          "fee": {
            "type": "string",
            "x-sdkwork-decimal-string": true
          },
          "currencyCode": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "orderItemId": {
            "type": "string"
          },
          "skuId": {
            "type": "string"
          },
          "driveSpaceId": {
            "type": "string"
          },
          "driveSpaceType": {
            "type": "string",
            "enum": [
              "notary"
            ]
          },
          "driveFolderNodeId": {
            "type": "string"
          },
          "parties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Party"
            }
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotaryDocument"
            }
          },
          "timeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelineEvent"
            }
          },
          "version": {
            "type": "string",
            "format": "int64",
            "x-sdkwork-int64-string": true
          }
        }
      },
      "NotaryCasePage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "pageInfo"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotaryCase"
            }
          },
          "pageInfo": {
            "$ref": "#/components/schemas/PageInfo"
          }
        }
      },
      "CreateNotaryCaseRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "skuId",
          "title",
          "applicantName"
        ],
        "properties": {
          "skuId": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "remarks": {
            "type": "string",
            "maxLength": 2000
          },
          "applicantName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "primaryNotaryMembershipId": {
            "type": "string",
            "description": "Organization member id selected as the primary notary for this case."
          },
          "parties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateNotaryPartyRequest"
            }
          },
          "initialDriveNodeIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "driveFolderName": {
            "type": "string",
            "maxLength": 255
          }
        }
      },
...
      "CreateNotaryPartyRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "role",
          "identityNo"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "role": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "partyType": {
            "type": "string",
            "enum": [
              "natural_person",
              "legal_person",
              "organization"
            ],
            "default": "natural_person"
          },
          "identityType": {
            "type": "string",
            "default": "resident_identity_card"
          },
          "identityNo": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "phone": {
            "type": "string",
            "maxLength": 64
          },
          "gender": {
            "type": "string",
            "maxLength": 32
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "address": {
            "type": "string",
            "maxLength": 500
          },
          "remarks": {
            "type": "string",
            "maxLength": 2000
          },
          "identityValidDateStart": {
            "type": "string",
            "format": "date"
          },
          "identityValidDateEnd": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "SdkWorkApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "enum": [
              0
            ],
            "default": 0,
            "minimum": 0,
            "maximum": 0
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
...
      "SdkWorkPlatformErrorCode": {
        "type": "integer",
        "format": "int32",
        "minimum": 40001,
        "maximum": 79999,
        "description": "Platform or domain error code per API_SPEC.md §15.3."
      },
...
    }
  }
}